+++ xFunc.Maths/Analyzers/IAnalyzer{TResult,TContext}.cs
@@ -370,6 +370,14 @@ public interface IAnalyzer<out TResult, in TContext>
     /// <param name="exp">The expression.</param>
     /// <param name="context">The context.</param>
     /// <returns>The result of analysis.</returns>
+    TResult Analyze(Curry exp, TContext context);
+
+    /// <summary>
+    /// Analyzes the specified expression.
+    /// </summary>
+    /// <param name="exp">The expression.</param>
+    /// <param name="context">The context.</param>
+    /// <returns>The result of analysis.</returns>
     TResult Analyze(Variable exp, TContext context);
 
     /// <summary>