+++ xFunc.Maths/Expressions/Lambda.cs
@@ -88,8 +88,29 @@ public readonly struct Lambda : IEquatable<Lambda>
     /// </summary>
     /// <param name="parameters">An object that contains all parameters and functions for expressions.</param>
     /// <returns>A result of the execution.</returns>
-    public object Call(ExpressionParameters parameters)
+    public object Call(ExpressionParameters? parameters)
         => Body.Execute(parameters);
+
+    /// <summary>
+    /// Calls the function.
+    /// </summary>
+    /// <param name="callParameters">The list of explicit parameters to call lambda with.</param>
+    /// <param name="expressionParameters">An object that contains all parameters and functions for expressions.</param>
+    /// <returns>A result of the execution.</returns>
+    /// <seealso cref="CallExpression"/>
+    public object Call(ImmutableArray<IExpression> callParameters, ExpressionParameters? expressionParameters)
+    {
+        var nestedScope = ExpressionParameters.CreateScoped(CapturedScope ?? expressionParameters);
+        var zip = Parameters.Zip(callParameters, (parameter, expression) => (parameter, expression));
+        foreach (var (parameter, expression) in zip)
+            nestedScope[parameter] = new ParameterValue(expression.Execute(expressionParameters));
+
+        var result = Call(nestedScope);
+        if (result is Lambda lambdaResult)
+            return lambdaResult.Capture(nestedScope);
+
+        return result;
+    }
 
     /// <summary>
     /// Returns a new lambda instance with captured parameters.
@@ -98,6 +119,26 @@ public readonly struct Lambda : IEquatable<Lambda>
     /// <returns>The lambda with captured parameters.</returns>
     public Lambda Capture(ExpressionParameters? parameters)
         => new Lambda(Parameters, Body, parameters);
+
+    /// <summary>
+    /// Converts the current lambda into a list of nested lambdas.
+    /// </summary>
+    /// <returns>The lambda.</returns>
+    /// <seealso cref="Expressions.Curry"/>
+    public Lambda Curry()
+    {
+        if (Parameters.Length <= 1)
+            return this;
+
+        var body = Body;
+
+        for (var i = Parameters.Length - 1; i > 0; i--)
+            body = new LambdaExpression(new Lambda(ImmutableArray.Create(Parameters[i]), body));
+
+        var lambda = new Lambda(ImmutableArray.Create(Parameters[0]), body);
+
+        return lambda;
+    }
 
     /// <summary>
     /// Converts <see cref="Lambda"/> to <see cref="LambdaExpression"/>.