+++ xFunc.Tests/Expressions/LambdaExpressionTest.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Dmytro Kyshchenko. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 namespace xFunc.Tests.Expressions;
 
 public class LambdaExpressionTest